extract tool selectors in multiple files
