+++ frontend/scripts/react-components/nav/filters-nav/recolor-by-selector/recolor-by-selector.js
@@ -1,6 +1,6 @@
 import { connect } from 'react-redux';
 import { selectRecolorBy } from 'react-components/tool/tool.actions';
-import { getToolRecolorGroups } from 'react-components/tool/tool.selectors';
+import { getToolRecolorGroups } from 'react-components/tool-links/tool-links.selectors';
 import RecolorBySelector from './recolor-by.component';
 import { getSelectedRecolorByValue, getRecolorByOptions } from './recolor-by-selector.selectors';
