+++ frontend/scripts/react-components/tool-links/tool-links.reducer.js
@@ -40,11 +40,11 @@ export const toolLinksInitialState = {
   currentQuant: null,
   detailedView: false,
   forcedOverview: false,
-  expandedNodesIds: [],
   highlightedNodeId: null,
   flowsLoading: false,
   selectedBiomeFilter: null,
   selectedColumnsIds: null,
+  expandedNodesIds: [],
   selectedNodesIds: [],
   selectedRecolorBy: null,
   selectedResizeBy: null,
