+++ frontend/scripts/react-components/tool/sankey/sankey.container.js
@@ -14,12 +14,12 @@ import {
   getIsReExpand,
   getHasExpandedNodesIds
 } from 'react-components/tool/sankey/sankey.selectors';
+import { getSelectedRecolorBy } from 'react-components/tool/tool.selectors';
 import {
   getVisibleNodesByColumn,
   getMergedLinks,
-  getNodesColored,
-  getSelectedRecolorBy
-} from 'react-components/tool/tool.selectors';
+  getNodesColored
+} from 'react-components/tool-links/tool-links.selectors';
 import Sankey from 'react-components/tool/sankey/sankey.component';
 
 const mapStateToProps = state => ({
